--- a/Misc/NEWS.d/next/Library/2025-02-13-02-03-38.gh-issue-118761.le_qEg.rst
+++ b/Misc/NEWS.d/next/Library/2025-02-13-02-03-38.gh-issue-118761.le_qEg.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`cmd` by lazy importing :mod:`inspect`. Patch by
+Semyon Moroz.