I use lazy import for `inspect` module (called 1 time).\
And I suggest getting rid of `import string`. This will improve import time as much as possible (benchmarks without `string` module separately in second comment).

### CPython configure flags:
```shell
./configure --enable-optimizations --with-lto --enable-loadable-sqlite-extensions
```

## Benchmarks:

Running: `pipx install tuna && ./python -X importtime -c 'import cmd' 2> import.log && tuna import.log`

### Total import time: 0.023s -> 0.013s = x1.77 as fast

| main branch | PR branch |
|-------------|-----------|
| ![Screenshot from 2025-02-13 01-59-59](https://github.com/user-attachments/assets/10b89c5c-882a-4452-b167-86912a114756) | ![Screenshot from 2025-02-13 01-54-43](https://github.com/user-attachments/assets/06f0f1c6-a538-46ad-9b19-30f8b4c012ff) |

### `cmd` import time: 0.016s -> 0.007s = x2.29 as fast

| main branch | PR branch |
|-------------|-----------|
| ![Screenshot from 2025-02-13 02-00-06](https://github.com/user-attachments/assets/55d0bd56-1a6b-415c-8661-bf3ba768e686) | ![Screenshot from 2025-02-13 01-54-50](https://github.com/user-attachments/assets/700f8116-a41a-495c-86f6-df596509b450) |

### hyperfine: 23.7ms -> 15.0ms = x1.58 as fast

### Main branch:
```shell
$ hyperfine --warmup 11 "./python -c 'import cmd'"
Benchmark 1: ./python -c 'import cmd'
  Time (mean ± σ):      23.7 ms ±   0.3 ms    [User: 20.3 ms, System: 3.3 ms]
  Range (min … max):    23.1 ms …  25.7 ms    123 runs
```

### PR branch:
```shell
$ hyperfine --warmup 11 "./python -c 'import cmd'"
Benchmark 1: ./python -c 'import cmd'
  Time (mean ± σ):      15.0 ms ±   0.3 ms    [User: 12.6 ms, System: 2.4 ms]
  Range (min … max):    14.4 ms …  16.0 ms    189 runs
```

<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->