@donBarbos 2025-02-12
I also suggest removing `import string` and making the following one change:
```python
# from this:
IDENTCHARS = string.ascii_letters + string.digits + '_'
# to this:
IDENTCHARS = ('ABCDEFGHIJKLMNOPQRSTUVWXYZ'
              'abcdefghijklmnopqrstuvwxyz'
              '0123456789'
              '_')
```
That's all!

## Benchmarks for `cmd` module without `import string`

comparing with the main branch here are the results:

### Total import time: 0.023s -> 0.008s = x2.88 as fast

| main branch | PR branch |
|-------------|-----------|
| ![Screenshot from 2025-02-13 01-59-59](https://github.com/user-attachments/assets/92f8ae14-e012-4fe2-a3e5-1cebcc4dd537) | ![Screenshot from 2025-02-13 01-53-26](https://github.com/user-attachments/assets/6ca4ff94-444e-483b-be41-80878a2aec8f) |

### `cmd` import time: 0.016s -> 0.001s = x16 as fast (or more :)

| main branch | PR branch |
|-------------|-----------|
| ![Screenshot from 2025-02-13 02-00-06](https://github.com/user-attachments/assets/bc63a1b2-e83a-428e-b496-91b457741447) | ![Screenshot from 2025-02-13 01-53-34](https://github.com/user-attachments/assets/9c173370-0279-4187-83c5-b0b0e6c70d28) |

### hyperfine: 23.7ms -> 9.0ms = x2.64 as fast
```shell
$ hyperfine --warmup 11 "./python -c 'import cmd'"
Benchmark 1: ./python -c 'import cmd'
  Time (mean ± σ):       9.0 ms ±   0.6 ms    [User: 7.6 ms, System: 1.4 ms]
  Range (min … max):     8.4 ms …  13.0 ms    288 runs
```